Create combined extractor
